Remove vbeta subpackage
