+++ instrumentation/opentelemetry-api-beta/src/main/java/io/opentelemetry/auto/instrumentation/opentelemetryapi/context/UnshadedScope.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.opentelemetry.auto.instrumentation.opentelemetryapi.context;
+
+import unshaded.io.opentelemetry.context.Scope;
+
+public class UnshadedScope implements Scope {
+
+  private final io.opentelemetry.context.Scope shadedScope;
+
+  public UnshadedScope(final io.opentelemetry.context.Scope shadedScope) {
+    this.shadedScope = shadedScope;
+  }
+
+  @Override
+  public void close() {
+    shadedScope.close();
+  }
+}
